ロボットリスト作成。フォーカスロボットをプロパティに登録。UI挙動を修正

+++ UnityProject/Assets/Scripts/AvatarComponents/AvatarSetting.cs
@@ -58,6 +58,6 @@ public class AvatarSetting : MonoBehaviourPunCallbacks
     {
         SetAvatarRoll();
         SetAvatarCamera();
-        sceneMaster.photonMng.AddToRoleDic(PhotonView.Get(this).ViewID, avatarRole);
+        sceneMaster.photonMng.AddToRoleDic(PhotonView.Get(this).ViewID, (int)avatarRole);
     }
 }
